requirements.txt
pytest
pyyaml
pandas
python-dotenv